apps/api/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.1
psycopg[binary]==3.1.18
argon2-cffi==23.1.0
gunicorn==21.2.0
redis==5.0.3
whitenoise==6.6.0
